Toma de datos del sensor y su cambio de estado

+++ Firmware_ESP32/ProyectoCNC/include/current_sensor.h
@@ -2,10 +2,10 @@
 #define SAMPLE_PERIOD_US 1000 // 1000us o 1ms de periodo de muestreo del ADC, para la LDR y LM35
 #include <stdbool.h>
 typedef struct {
-    int s_I; // corriente spindle
-    int x_I;  
-    int y_I;  
-    int z_I; //corriente motor Z
+    float s_I; // corriente spindle
+    float x_I;  
+    float y_I;  
+    float z_I; //corriente motor Z
 } consumo_cnc_t;
 void I_sensor_init(void);
 bool read_I_sensor(consumo_cnc_t *data);
